brain/requirements.txt
aiohttp==2.3.10
async-timeout==2.0.1
awscli==1.11.13
botocore==1.4.70
Box2D==2.3.2
chardet==2.3.0
Click==7.0
colorama==0.3.7
command-not-found==0.3
defer==1.0.6
dnspython==1.16.0
docutils==0.12
Flask==1.1.1
Flask-SocketIO==4.2.1
gevent==1.4.0
greenlet==0.4.15
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.0
language-selector==0.1
lightdm-gtk-greeter-settings==1.2.1
lxml==3.5.0
MarkupSafe==1.1.1
monotonic==1.5
multidict==4.7.4
netifaces==0.10.6
paho-mqtt==1.5.0
Pillow==3.1.2
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
Pygments==2.1
pygobject==3.20.0
python-apt==1.1.0b1+ubuntu0.16.4.8
python-dateutil==2.4.2
python-debian==0.1.27
python-engineio==3.11.2
python-socketio==4.4.0
requests==2.9.1
roman==2.0.0
rsa==3.2.3
s3transfer==0.1.9
six==1.10.0
system-service==0.3
ubuntu-drivers-common==0.0.0
ufw==0.35
urllib3==1.13.1
usb-creator==0.3.0
websocket-client==0.57.0
Werkzeug==1.0.0
xkit==0.0.0
yarl==1.4.2
